refactor: give the bit-level primitives a home in composer::bits
`recompose_bits` sat in `composer.rs`, which never calls it;
`component_decomposition` was the last `component_*` gadget left there;
and `component_boolean` had landed in `select`, where none of the
selection gadgets calls it — their callers are `range_check`,
`component_decomposition` and `component_select_identity`, in three
other modules. All three now sit in a `bits` submodule, so the gadget
modules build on a named primitive layer rather than on each other.

`recompose_bits` becomes `pub(super)`, which is the reach it already
had as a private item of `composer`.
