+++ src/composer/truncate.rs
@@ -9,7 +9,8 @@
 
 use dusk_bls12_381::BlsScalar;
 
-use super::{Composer, Constraint, Witness, recompose_bits};
+use super::bits::recompose_bits;
+use super::{Composer, Constraint, Witness};
 
 #[cfg(test)]
 mod truncate_soundness_tests;
